resolved conflicts

+++ lib/components/alert_dialog.dart
@@ -5,12 +5,18 @@ import '../provider/songItem.dart';
 class ConfirmDeleteAlert extends StatelessWidget {
   const ConfirmDeleteAlert({
     Key key,
-    @required this.deleteFunction,
+    this.deleteFunction,
     @required this.songProvider,
+    this.playListDelete,
+    this.playListDeleteSingle,
+    this.playlistName,
   }) : super(key: key);
 
   final void Function() deleteFunction;
+  final void Function(String, List<String>) playListDelete;
+  final void Function(String, String, List<String>) playListDeleteSingle;
   final SongProvider songProvider;
+  final String playlistName;
 
   @override
   Widget build(BuildContext context) {
@@ -20,15 +26,23 @@ class ConfirmDeleteAlert extends StatelessWidget {
       actions: [
         FlatButton(
           child: Text("No"),
-          onPressed: () => Navigator.of(context).pop(),
+          onPressed: () => Navigator.of(context).pop(false),
         ),
         FlatButton(
           child: Text("Yes"),
           onPressed: () {
-            deleteFunction();
+            if (deleteFunction != null) {
+              deleteFunction();
+            } else if (playListDeleteSingle != null) {
+              playListDeleteSingle(
+                  "playLists", playlistName, songProvider.queuePath);
+            } else {
+              playListDelete("playLists", songProvider.keys);
+            }
             songProvider.changeBottomBar(false);
             songProvider.setQueueToNull();
-            Navigator.of(context).pop();
+            songProvider.setKeysToNull();
+            Navigator.of(context).pop(true);
           },
         ),
       ],
